+++ tests/externallib_test.php
@@ -111,7 +111,9 @@ final class externallib_test extends \advanced_testcase {
             $this->generator->teardown();
             $this->generator = null;
         }
-        $this->engine->delete('core_mocksearch-mock_search_area');
+        if ($this->engine) {
+            $this->engine->delete('core_mocksearch-mock_search_area');
+        }
         parent::tearDown();
     }
