+++ tests/local/chunking/chunking_engine_test.php
@@ -126,8 +126,10 @@ final class chunking_engine_test extends advanced_testcase {
             $this->generator->teardown();
             $this->generator = null;
         }
-        $this->engine->delete();
-        sleep(1);
+        if ($this->engine) {
+            $this->engine->delete();
+            sleep(1);
+        }
         parent::tearDown();
     }
